클래스 구조 정리
타입 변환 경고 해결

+++ Universe2D/Application.h
@@ -28,7 +28,7 @@ private:
     int m_NowTime;
     int m_FrameCount;
 
-    float m_DeltaTime;
+    int m_DeltaTime;
 
 public:
     Application();
